+++ pyproject.toml
@@ -157,10 +157,6 @@ ignore = [
   # Allow use of non-cryptographic `random` module.
   "S311"
 ]
-unfixable = [
-  # Don't touch unused imports
-  "F401",
-]
 
 [tool.ruff.lint.flake8-tidy-imports]
 ban-relative-imports = "all"
